fix: check if brand is referenced in product before deleting

+++ src/main/java/com/musinsa/coord/brand/BrandService.java
@@ -5,6 +5,7 @@ import com.musinsa.coord.brand.dto.BrandResponse;
 import com.musinsa.coord.brand.dto.BrandUpdateRequest;
 import com.musinsa.coord.common.error.exception.BusinessException;
 import com.musinsa.coord.common.error.exception.ErrorCode;
+import com.musinsa.coord.product.ProductRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -14,9 +15,11 @@ import org.springframework.transaction.annotation.Transactional;
 public class BrandService {
 
     private BrandRepository brandRepository;
+    private ProductRepository productRepository;
 
-    public BrandService(BrandRepository brandRepository) {
+    public BrandService(BrandRepository brandRepository, ProductRepository productRepository) {
         this.brandRepository = brandRepository;
+        this.productRepository = productRepository;
     }
 
     public BrandResponse findBrand(Long id) {
@@ -60,6 +63,10 @@ public class BrandService {
     public void deleteBrand(Long id) {
         if (!brandRepository.existsById(id)) {
             throw new BusinessException(ErrorCode.RESOURCE_NOT_FOUND);
+        }
+        boolean hasAssociatedProducts = productRepository.existsByBrandId(id);
+        if (hasAssociatedProducts) {
+            throw new BusinessException(ErrorCode.BRAND_HAS_ASSOCIATED_PRODUCTS);
         }
         brandRepository.deleteById(id);
     }
